+++ src/canvas/widgets/disk_table.rs
@@ -148,7 +148,7 @@ impl DiskTableWidget for Painter {
                             .map(|calculated_width| Constraint::Length(*calculated_width as u16))
                             .collect::<Vec<_>>()),
                     )
-                    .header_gap(0),
+                    .header_gap(app_state.app_config_fields.table_gap),
                 margined_draw_loc[0],
             );
         }